ARC-3613: Move file icon to helper function and use for conversation and documents
