+++ src/modules/account/components/MaterialRequestDetailBlade/MaterialRequestConversationMessage.tsx
@@ -11,6 +11,7 @@ import { Button } from '@meemoo/react-components';
 import Html from '@shared/components/Html/Html';
 import { Icon } from '@shared/components/Icon';
 import { IconNamesLight } from '@shared/components/Icon/Icon.enums';
+import { getFileNameIcon } from '@shared/helpers/get-file-name-icon';
 import { tHtml, tText } from '@shared/helpers/translate';
 import { asDate, formatLongDate, formatMediumDateWithTime } from '@shared/utils/dates';
 import clsx from 'clsx';
@@ -237,30 +238,6 @@ export const MaterialRequestConversationMessage: FC<MaterialRequestConversationM
 		);
 	};
 
-	const determineAttachmentIcon = (fileName: string) => {
-		const extension = fileName.split('.').pop()?.toLowerCase();
-
-		switch (extension) {
-			case 'pdf':
-				return IconNamesLight.FilePdf;
-			case 'doc':
-			case 'docx':
-				return IconNamesLight.FileDoc;
-			case 'xls':
-			case 'xlsx':
-				return IconNamesLight.FileXls;
-			case 'jpg':
-			case 'jpeg':
-				return IconNamesLight.FileJpg;
-			case 'png':
-				return IconNamesLight.FilePng;
-			case 'csv':
-				return IconNamesLight.FileCsv;
-			default:
-				return IconNamesLight.File;
-		}
-	};
-
 	return (
 		<div
 			className={clsx(
@@ -285,7 +262,7 @@ export const MaterialRequestConversationMessage: FC<MaterialRequestConversationM
 						passHref
 						className={clsx(styles['p-conversation-messages__message__attachment'])}
 					>
-						<Icon name={determineAttachmentIcon(attachment.attachmentFilename)}></Icon>
+						<Icon name={getFileNameIcon(attachment.attachmentFilename)}></Icon>
 						{attachment.attachmentFilename}
 					</Link>
 				))}
